chore: move actions, annotations and groups UI into their products
Per /placing-product-frontend-code, a scene dir routes to the product of
the same name. All three products already existed, so this is placement
only: no behavior changes, and no CI run gets faster.

Annotations and groups register their scenes and routes in their own
manifests and drop out of appScenes and scenes.ts; annotations also moves
its two URL helpers, while groups already owned its URLs. The Scene enum
entries stay, which is how already-migrated products are wired. The
annotations barrel index.ts had no consumers and is deleted, not moved.

Moving a file into a product package means that package has to declare
what it imports, so annotations and groups gain the dependencies their
files use. Kea type-only imports take the deep relative path into
frontend/src, matching the actions product.

Groups has 15 importers outside its directory, within the range already
accepted for funnels (25). Core now reaches into the product on those
edits, as it does for funnels and retention.

Persons is deliberately left alone: PersonDisplay has 44 importers and
person-utils 20, so at 78 external importers it raises the shared-hub
question that trends is already blocked on, rather than being placement.
