+++ frontend/src/lib/components/ActivityLog/describers.tsx
@@ -3,8 +3,6 @@ import { ActivityDescriber as errorTrackingActivityDescriber } from '@posthog/pr
 import { instanceSettingActivityDescriber } from 'lib/components/ActivityLog/activityDescriptions/instanceSettingActivityDescriber'
 import { tagActivityDescriber } from 'lib/components/ActivityLog/activityDescriptions/tagActivityDescriber'
 import { ActivityLogItem, Describer, defaultDescriber } from 'lib/components/ActivityLog/humanizeActivity'
-import { actionActivityDescriber } from 'scenes/actions/actionActivityDescriber'
-import { annotationActivityDescriber } from 'scenes/annotations/activityDescriptions'
 import { userActivityDescriber } from 'scenes/authentication/shared/activityDescriptions'
 import { billingActivityDescriber } from 'scenes/billing/activityDescriptions'
 import { cohortActivityDescriber } from 'scenes/cohorts/activityDescriptions'
@@ -16,7 +14,6 @@ import { dataWarehouseSavedQueryActivityDescriber } from 'scenes/data-warehouse/
 import { experimentActivityDescriber } from 'scenes/experiments/experimentActivityDescriber'
 import { exportedAssetActivityDescriber } from 'scenes/exports/activityDescriptions'
 import { flagActivityDescriber } from 'scenes/feature-flags/activityDescriptions'
-import { groupActivityDescriber } from 'scenes/groups/activityDescriptions'
 import { hogFunctionActivityDescriber } from 'scenes/hog-functions/misc/activityDescriptions'
 import { notebookActivityDescriber } from 'scenes/notebooks/Notebook/notebookActivityDescriber'
 import { personActivityDescriber } from 'scenes/persons/activityDescriptions'
@@ -38,13 +35,16 @@ import { teamActivityDescriber } from 'scenes/team-activity/teamActivityDescribe
 
 import { ActivityScope } from '~/types'
 
+import { actionActivityDescriber } from 'products/actions/frontend/actionActivityDescriber'
 import { promptActivityDescriber } from 'products/ai_observability/frontend/prompts/promptActivityDescriber'
 import { promptLabelActivityDescriber } from 'products/ai_observability/frontend/prompts/promptLabelActivityDescriber'
 import { alertConfigurationActivityDescriber } from 'products/alerts/frontend/components/activityDescriptions'
+import { annotationActivityDescriber } from 'products/annotations/frontend/activityDescriptions'
 import { canvasActivityDescriber } from 'products/canvas/frontend/activityDescriber'
 import { ticketActivityDescriber } from 'products/conversations/frontend/activityDescriber'
 import { externalDataSourceActivityDescriber } from 'products/data_warehouse/frontend/shared/components/activityDescriptions'
 import { endpointActivityDescriber } from 'products/endpoints/frontend/activityDescriber'
+import { groupActivityDescriber } from 'products/groups/frontend/activityDescriptions'
 import { signalScoutConfigActivityDescriber } from 'products/signals/frontend/activityDescriber'
 import { stamphogRepoConfigActivityDescriber } from 'products/stamphog/frontend/activityDescriber'
 import { workflowActivityDescriber } from 'products/workflows/frontend/Workflows/misc/workflowActivityDescriber'
