--- frontend/src/scenes/annotations/index.ts
@@ -1,1 +1,0 @@
-export { Annotations } from './Annotations'
