+++ frontend/src/scenes/appScenes.ts
@@ -7,7 +7,6 @@ export const appScenes: Record<Scene | string, () => any> = {
     ...productScenes,
     [Scene.AdvancedActivityLogs]: () => import('./audit-logs/AdvancedActivityLogsScene'),
     [Scene.AgenticAuthorize]: () => import('./agentic/AgenticAuthorize'),
-    [Scene.Annotations]: () => import('./annotations/Annotations'),
     [Scene.Approval]: () => import('./approvals/ApprovalDetail'),
     [Scene.AsyncMigrations]: () => import('./instance/AsyncMigrations/AsyncMigrations'),
     [Scene.MaterializedColumns]: () => import('./data-management/MaterializedColumns/MaterializedColumns'),
@@ -44,9 +43,6 @@ export const appScenes: Record<Scene | string, () => any> = {
     [Scene.ExploreSessions]: () => import('./activity/explore/SessionsScene'),
     [Scene.FeatureFlag]: () => import('./feature-flags/FeatureFlag'),
     [Scene.FeatureFlags]: () => import('./feature-flags/FeatureFlags'),
-    [Scene.Group]: () => import('./groups/Group'),
-    [Scene.GroupsNew]: () => import('./groups/GroupsNew'),
-    [Scene.Groups]: () => import('./groups/Groups'),
     [Scene.HogFunction]: () => import('./hog-functions/HogFunctionScene'),
     [Scene.Insight]: () => import('./insights/InsightScene'),
     [Scene.InsightQuickStart]: () => import('./insights/InsightQuickStart/InsightQuickStart'),
