+++ frontend/src/scenes/persons/PersonScene.tsx
@@ -25,7 +25,6 @@ import { copyToClipboard } from 'lib/utils/copyToClipboard'
 import { isMobile } from 'lib/utils/dom'
 import { pluralize } from 'lib/utils/strings'
 import { tryDecodeURIComponent } from 'lib/utils/url'
-import { RelatedGroups } from 'scenes/groups/RelatedGroups'
 import { NotebookSelectButton } from 'scenes/notebooks/NotebookSelectButton/NotebookSelectButton'
 import { NotebookNodeType } from 'scenes/notebooks/types'
 import { PersonDeleteModal } from 'scenes/persons/PersonDeleteModal'
@@ -53,6 +52,7 @@ import {
 
 import { ComposeTicketButton } from 'products/conversations/frontend/components/ComposeTicket'
 import { FeedbackButton } from 'products/customer_analytics/frontend/components/FeedbackButton'
+import { RelatedGroups } from 'products/groups/frontend/components/RelatedGroups'
 
 import { MergeSplitPerson } from './MergeSplitPerson'
 import { asDisplay, pickBestPersonDistinctId } from './person-utils'
