+++ products/actions/frontend/actionActivityDescriber.tsx
