+++ products/annotations/frontend/activityDescriptions.tsx
