+++ products/annotations/frontend/pages/Annotations.stories.tsx
