+++ products/annotations/frontend/pages/__mocks__/annotations.json
