+++ products/annotations/manifest.tsx
@@ -3,19 +3,28 @@
  *
  * Defines scenes, routes, URLs, and navigation for this product.
  */
-import { ProductManifest } from '../../frontend/src/types'
+import { AnnotationType, ProductManifest } from '../../frontend/src/types'
 
 export const manifest: ProductManifest = {
     name: 'Annotations',
     scenes: {
-        // Define scenes here
+        Annotations: {
+            name: 'Annotations',
+            import: () => import('./frontend/pages/Annotations'),
+            projectBased: true,
+            description:
+                'Annotations allow you to mark when certain changes happened so you can easily see how they impacted your metrics.',
+            iconType: 'annotation',
+        },
     },
     routes: {
-        // Define routes here
+        '/data-management/annotations': ['Annotations', 'annotations'],
+        '/data-management/annotations/:id': ['Annotations', 'annotation'],
     },
     redirects: {},
     urls: {
-        // Define URL helpers here
+        annotations: (): string => '/data-management/annotations',
+        annotation: (id: AnnotationType['id'] | ':id'): string => `/data-management/annotations/${id}`,
     },
     fileSystemTypes: {},
     treeItemsNew: [],
