+++ products/annotations/package.json
@@ -4,7 +4,9 @@
         "backend:test": "pytest -c ../../pytest.ini --rootdir ../.. backend/ -v --tb=short"
     },
     "dependencies": {
-        "kea": "catalog:"
+        "kea": "catalog:",
+        "kea-forms": "catalog:",
+        "kea-router": "catalog:"
     },
     "devDependencies": {
         "@storybook/react": "catalog:",
@@ -13,6 +15,8 @@
     "peerDependencies": {
         "@posthog/brand": "catalog:",
         "@posthog/icons": "catalog:",
+        "@testing-library/dom": "*",
+        "@testing-library/user-event": "*",
         "@types/react": "*",
         "clsx": "*",
         "react": "*"
