+++ products/groups/frontend/activityDescriptions.tsx
