+++ products/groups/frontend/cards/GroupNotebookCard.tsx
