+++ products/groups/frontend/components/GroupsIntroduction.tsx
