+++ products/groups/frontend/logics/relatedGroupsLogic.ts
@@ -7,7 +7,7 @@ import { teamLogic } from 'scenes/teamLogic'
 
 import { ActorType } from '~/types'
 
-import type { PersonActorType } from '../../types'
+import type { PersonActorType } from '../../../../frontend/src/types'
 
 // Generated by kea-typegen. Update if you're an agent, ignore if you're human.
 export interface relatedGroupsLogicValues {
